+++ level.py
@@ -6,6 +6,8 @@ from debug import debug
 from support import *
 from random import choice
 from weapon import Weapon
+from ui import UI
+
 class Level:
     def __init__(self):
         # sprite group setup
@@ -18,6 +20,9 @@ class Level:
 
         # sprite setup
         self.create_map()
+
+        # user interface
+        self.ui = UI()
 
 
     def create_map(self):
@@ -65,6 +70,8 @@ class Level:
         self.visible_sprites.custom_draw(self.player)
         self.visible_sprites.update()
         debug(self.player.status)
+
+        self.ui.display(self.player)
 
 class YSortCameraGroup(pygame.sprite.Group):
     def __init__(self):